refactor: extract superset_config.py builder; move ConfigMap build into controller/build
Extracts the superset_config.py rendering (TALISMAN default + Superset/OPA
settings + per-rolegroup properties + FILE_HEADER/FILE_FOOTER + Flask writer +
celery connection class) into config/superset_config::build(), and relocates the
ConfigMap assembler to controller/build/config_map.rs — matching the
config/<file-builder> + controller/build/config_map.rs layout of airflow, hdfs,
kafka and trino. The FILE_HEADER/FILE_FOOTER constants move into the builder,
removing the last stackable_operator::product_config_utils import. No behaviour
change (33 tests pass).
