+++ rust/operator-binary/src/config/mod.rs
@@ -1,4 +1,5 @@
 pub mod commands;
 pub mod product_logging;
 pub mod superset;
+pub mod superset_config;
 pub mod writer;
